trabalho hash matheus s/requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
Werkzeug==2.3.7
